common

+++ thinking-in-shardingsphere-jdbc/src/main/java/com/jpeony/sharding/jdbc/mapper/UserMapper.java
@@ -16,5 +16,5 @@ public interface UserMapper {
 
     List<User> queryAllUser();
 
-    User queryUserById(Long id);
+    User queryUserById(int id);
 }
